+++ Game/AI/Decks/ApophisExecutor.cs
@@ -604,7 +604,7 @@ namespace WindBot.Game.AI.Decks
                 .Max(card => (int?)card.GetDefensePower()) ?? -1;
         }
 
-        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, int hint, bool cancelable)
+        public override IList<ClientCard> OnSelectCard(IList<ClientCard> cards, int min, int max, long hint, bool cancelable)
         {
             ChainInfo currentChain = Duel.GetCurrentSolvingChainInfo();
             if (currentChain == null) {
@@ -1269,7 +1269,7 @@ namespace WindBot.Game.AI.Decks
             return base.OnSelectOption(options);
         }
 
-        public override int OnSelectPlace(int cardId, int player, CardLocation location, int available)
+        public override int OnSelectPlace(long cardId, int player, CardLocation location, int available)
         {
             if (player == 0 && location == CardLocation.MonsterZone)
             {
@@ -1336,7 +1336,7 @@ namespace WindBot.Game.AI.Decks
             return base.OnSelectPlace(cardId, player, location, available);
         }
 
-        public override bool OnSelectYesNo(int desc)
+        public override bool OnSelectYesNo(long desc)
         {
             if (desc == Util.GetStringId(CardId.PrimiteDrillbeam, 0))
             {
